remove unnecessary web workers to prevent performance problems
